EventLoop.iteration(): avoid busy waiting
In order to avoid blocking forever when may_block is True (the
default), callers must be careful to ensure that at least one of the
following conditions is met:
	1) An event source or timeout is registered which is guaranteed
		to trigger at least on event (a call to an idle function
		only counts as an event if it returns a False value which
		causes it to be stop being called)
	2) Another thread is guaranteed to call one of the thread-safe
		methods which notify iteration to stop waiting (such as
		idle_add or timeout_add).
These rules ensure that iteration is able to block until an event
arrives, without doing any busy waiting that would waste CPU time.

This will fix busy waiting which would be triggered by
PopenPipeBlockingIOTestCase when waiting for the thread from
PipeReaderBlockingIO to call idle_add.