install_qa_check_macho: allow relative install_names
relative install_names (@..../) will be used in binaries a lot, and if
they are provided as such, there is nothing wrong with it.  Since they
disallow us to do any checks, we just do the least check we can do,
which is for the install_name self-reference.  We want to find the
library that points to in the install image, if not, it must be wrong.