Fix crash in depgraph._remove_pkg
Make sure we don't try to remove a packages twice. This happened in the
past when a package was its own digraph-child.

One way to trigger this is a DEPEND on itself as in the test case.

This bug was reported by slyfox on IRC. In his case the cause for the
self-edge is unknown. Unfortunately the system state is lost.