ebuild-helpers/xattr/install: use install-xattr
Currently bin/ebuild-helpers/xattr/install uses ${PORTAGE_BIN_PATH}/install.py
as a wrapper to coreutils' install to preserve a file's extended attributes when
installing, usually during src_install().  This is needed, for instance, when
preserving xattr based PaX flags, bug #465000.  However the python wrapper is
very slow, comment #42 of bug #465000.  A C wrapper was developed and tested,
bugs #501534 and #511984.  This patch checks for the existence of the C wrapper,
and uses it, falling back on the python wrapper only if not found, or if over-
ridden by ${PORTAGE_INSTALL_XATTR_IMPLEMENTATION}.