Repoman: Refactor repos.RepoSettings class, fix the repoman_settings not being re-assigned
When the repo is not in the confiig'd repos, it creates a ne repoman_settings object which contains
the temporaily activated repo.  main.py was not re-assigning repoman_settings to the new object.
Split out _add_repo() to it's own function.
Split out the gpg-sign specific function for git into plugin-ready functions, for later.
Detect the lack of a vcs type in vcs_settings to not do a gpg-sign call.