sets: Introduce @changed-deps to update packages which need dep changes
The @changed-deps set compares the RDEPEND and PDEPEND entries of
installed packages with the metadata of matching ebuilds, triggering
rebuilds whenever a difference in dependencies is detected.

This set can be used to force installed package metadata updates when
ebuild dependencies are changed in-place. In particular, it can be used
to restore dependency tree integrity after disabling --dynamic-deps.

Acked-by: Alexander Berntsen <bernalex@gentoo.org>
Acked-by: Brian Dolbec <dolsen@gentoo.org>