vdb: Preserve := when storing uninstalled dependency atoms
Make the slot operator processing code not remove the := operator from
vdb dependency entries when the underlying package is not installed.

This fixes a bug where dependencies like || ( A:= B:= ) were stored as
|| ( A:0/1= B ) if B was not installed. This especially confused
dependency comparisons in @changed-deps.

With this patch, Portage stores the original := (or :slot=) atom in the
vdb. This fixes the issue while preserving the same behavior in the
current dependency resolver.

Acked-by: Alexander Berntsen <bernalex@gentoo.org>
Acked-by: Brian Dolbec <dolsen@gentoo.org>
Reviewed-by: Zac Medico <zmedico@gentoo.org>