emaint/main.py: Change func to long_action for module action check
Due to the change making it possible for the --action to be different than the actual function
beign called.  func no longer was the same as the avalable actions listed.
Reverts the change made in:
Commit: e166eb68e838389505f15426a2f0011dd95be342
Author: Brian Dolbec <dolsen@gentoo.org> (Wed 12 Mar 2014 08:40:49 PM PDT)
Subject: portage/emaint/main.py: Fix a bug in the options parsing.