_quickpkg_dblink: fix bug #523152
This fixes the unmerge-backup and downgrade-backup features to be
compatible with the new setup.py quickpkg install location, while
preserving compatibility for running from a source tree (for unit tests
or developement purposes). If not running from a source tree, then the
PATH variable is used to locate the quickpkg binary.

X-Gentoo-Bug: 523152
X-Gentoo-Bug-URL: https://bugs.gentoo.org/show_bug.cgi?id=523152