AbstractPollTask._read_buf: read regardless of event flags (531724)
PipeReaderPtyTestCase shows that data may be lost unless we attempt to
read data for every poll event, regardless of the event flags.

Therefore, always read, regardless of the event flags. This is safe to
do because all consumers of this API use non-blocking mode and properly
handle EAGAIN (signaled when this method returns None).

X-Gentoo-Bug: 531724
X-Gentoo-Bug-URL: https://bugs.gentoo.org/show_bug.cgi?id=531724
Acked-by: Alexander Berntsen <bernalex@gentoo.org>