Add --sync-submodule <glsa|news|profiles> (534070)
This adds support for a new --sync-submodule option to both emerge and
emaint. When this option is used with the sync action, only the selected
submodules are synced. Each submodule is referenced using an abstract
identifier, which serves to hide the implementation details involving
the precise locations of specific submodules within each repository.

Currently, --sync-submodule has no effect for sync protocols other than
rsync, but the new SyncBase._get_submodule_paths() method will be useful
for implementing support in other SyncBase subclasses.

X-Gentoo-Bug: 534070
X-Gentoo-Bug-URL: https://bugs.gentoo.org/show_bug.cgi?id=534070
Acked-by: Alexander Berntsen <bernalex@gentoo.org>