Scheduler: increase visiblity of postinst failures (bug 547778)
This makes postinst failures behave more like "real" failures in
terms of output and the final emerge return code. However, in other
respects, these packages will be treated as successful installations.
This means that they will not cause emerge to immediately exit, and
they will not trigger recalculation of dependencies when --keep-going
is enabled.

X-Gentoo-Bug: 547778
X-Gentoo-Bug-URL: https://bugs.gentoo.org/show_bug.cgi?id=547778
Acked-by: Brian Dolbec <dolsen@gentoo.org>